Avoid IFD in the module definition.
Previously required that the package was built to evaluate
`defaultPatterns`. The `defaultPatterns` are now always symlinked into
the final patterns directory that the service sees. If a pattern is to
ve removed, an attribute with the same name as the default pattern file
can be set in defaultPatterns pointing to an empty file can be used to
have the same effect.